chore: fix attestation tests
