+++ x/evm/keeper/testdata/slc-payload.hex
@@ -1,0 +1,1 @@
+0xaa
